--- lib/src/core/base/eden_base_list_widget.dart
@@ -1,15 +1,0 @@
-import 'package:flutter_eden/eden.dart';
-import 'mixin/eden_refreshable_mixin.dart';
-
-abstract class EdenBaseListWidget<T extends EdenBaseListController>
-    extends EdenBaseWidget<T> with EdenRefreshableMixin<T> {
-  EdenBaseListWidget({super.key});
-
-  @override
-  Widget buildListView(BuildContext context, T controller);
-
-  @override
-  Widget buildBody(BuildContext context, T controller) {
-    return buildRefreshableList(context, controller);
-  }
-}
